Separate slot related functions
